+++ Sample/PhysicallyBasedRendering/main.cpp
@@ -22,16 +22,8 @@ int main(int ac, char** av)
 {
 	try
 	{
-		// FIXME(anirul): Fix this by using the command line or the environment
-		// variables and pass the parameters to the application.
-		Application app(
-			sgl::CreateSDLOpenGL({ 640, 480 }), 
-			Application::draw_model_enum::SPHERE, 
-			Application::texture_model_enum::METAL);
-		if (!app.Startup())
-		{
-			return -1;
-		}
+		Application app(sgl::CreateSDLOpenGL({ 640, 480 }));
+		app.Startup();
 		app.Run();
 	}
 	catch (std::exception ex)
