+++ Sample/Simple/Application.h
@@ -8,14 +8,9 @@ class Application
 {
 public:
 	Application(const std::shared_ptr<sgl::WindowInterface>& window);
-	bool Startup();
+	void Startup();
 	void Run();
 
 protected:
-	std::shared_ptr<sgl::Mesh> CreateAppleMesh(
-		const std::shared_ptr<sgl::Device>& device) const;
-	std::shared_ptr<sgl::Mesh> CreateCubeMapMesh(
-		const std::shared_ptr<sgl::Device>& device,
-		const std::shared_ptr<sgl::TextureCubeMap>& texture) const;
 	std::shared_ptr<sgl::WindowInterface> window_;
 };
