+++ Sample/Simple/main.cpp
@@ -23,10 +23,7 @@ int main(int ac, char** av)
 	try
 	{
 		Application app(sgl::CreateSDLOpenGL({ 640, 480 }));
-		if (!app.Startup())
-		{
-			return -1;
-		}
+		app.Startup();
 		app.Run();
 	}
 	catch (std::exception ex)
